fix(web): scope share-modal pending count to students; rename helper to acceptShareSummary
Code review found the share-modal reach count added an all-roles pending
count (counts.pending, which includes staff-team invites) to a
student-only enrolled count, so a lone pending co-teacher/TA read as '1
student can accept' and silenced the no-students warning — the #376 case
the feature guards. Derive a student-scoped pending count from the team
rows via hasStudentEnrollment. Also rename acceptShareWarning.{ts,test.ts}
to acceptShareSummary.* to match the exported AcceptShareSummary/
resolveAcceptShareSummary API.
