+++ web/src/pages/submissions/acceptShareSummary.ts
